neural network code (python)
it contains python code for the neural network. This neural network works with quadratic cost function and backpropagation. The number of neurons in each layer can be entered by the user during instantiating the object of the class. Except the first and the last one which contains 784 and 10  neurons respectively.